+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/get/GetFile.kt
@@ -3,7 +3,7 @@ package com.github.insanusmokrassar.TelegramBotAPI.requests.get
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.PathedFile
-import com.github.insanusmokrassar.TelegramBotAPI.types.files.abstracts.fileIdField
+import com.github.insanusmokrassar.TelegramBotAPI.types.fileIdField
 import kotlinx.serialization.*
 
 @Serializable