+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/files/PhotoSize.kt
@@ -1,6 +1,8 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.files
 
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
+import com.github.insanusmokrassar.TelegramBotAPI.types.fileIdField
+import com.github.insanusmokrassar.TelegramBotAPI.types.fileSizeField
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.abstracts.*
 import kotlinx.serialization.*
 import kotlinx.serialization.internal.ArrayListSerializer