+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/files/abstracts/TelegramMediaFile.kt
@@ -2,10 +2,6 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.files.abstracts
 
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
 
-internal const val fileIdField = "file_id"
-internal const val fileSizeField = "file_size"
-internal const val filePathField = "file_path"
-
 /**
  * Declare common part of media files in Telegram. Note: it is not representation of `File` type
  */